requirements.txt
freezegun==1.2.2
pytest==7.2.0
pytest-django==4.5.2
pytest-cov==4.0.0
django==4.1.3
django-rest-framework==0.1.0
django-environ==0.9.0
django-cors-headers==3.13.0
psycopg2==2.9.5
Pillow==9.3.0
